+++ options/defaults.lua
@@ -572,6 +572,9 @@ SAO.defaults = {
                 [63167] = { -- Decimation (Wrath+)
                     [0] = true,
                 },
+                [440873] = { -- Decimation (Season of Discovery)
+                    [0] = true,
+                },
                 [47283] = { -- Empowered Imp (Wrath+)
                     [0] = true,
                 },
@@ -580,7 +583,7 @@ SAO.defaults = {
                 },
             },
             glow = {
-                [1120] = { -- Drain Soul (Wrath+)
+                [1120] = { -- Drain Soul (Season of Discovery, Wrath+)
                     [1120] = "spec:1", -- Drain Soul
                 },
                 [17877] = { -- Shadowburn
@@ -588,12 +591,16 @@ SAO.defaults = {
                 },
                 [17941] = { -- Nightfall
                     [686] = true, -- Shadow Bolt
+                    [403841] = true, -- Shadow Cleave (Season of Discovery)
                 },
                 [71165] = { -- Molten Core (Wrath+)
                     [29722] = true, -- Incinerate
                     [6353]  = true, -- Soul Fire (not for Cata)
                 },
                 [63167] = { -- Decimation (Wrath+)
+                    [6353] = true, -- Soul Fire
+                },
+                [440873] = { -- Decimation (Season of Discovery)
                     [6353] = true, -- Soul Fire
                 },
                 [54277] = { -- Backdraft (Cataclysm)